+++ media_softlet/agnostic/Xe3_M_plus/Xe3_LPM_base/codec/hal/enc/hevc/features/encode_hevc_vdenc_const_settings_xe3_lpm_base_open.h
@@ -187,8 +187,8 @@ setting->vdencCmd2Settings.emplace_back(
                     return MOS_STATUS_SUCCESS;
                 }
 
-                uint8_t tmp0 = 0;
-                uint8_t tmp1 = 0;
+                uint32_t tmp0 = 0;
+                uint32_t tmp1 = 0;
 
                 if (m_hevcPicParams->CustomRoundingOffsetsParams.fields.EnableCustomRoudingIntra)
                 {
