+++ media_softlet/agnostic/Xe_M_plus/Xe_LPM_plus_base/codec/hal/enc/hevc/features/encode_hevc_vdenc_const_settings_xe_lpm_plus_base.cpp
@@ -508,8 +508,8 @@ MOS_STATUS EncodeHevcVdencConstSettingsXe_Lpm_Plus_Base::SetVdencCmd2Settings()
                         return MOS_STATUS_SUCCESS;
                     }
 
-                    uint8_t tmp0 = 0;
-                    uint8_t tmp1 = 0;
+                    uint32_t tmp0 = 0;
+                    uint32_t tmp1 = 0;
 
                     if (m_hevcPicParams->CustomRoundingOffsetsParams.fields.EnableCustomRoudingIntra)
                     {
